# openapi.yaml
openapi: 3.0.3
info:
  title: Trine Empire Core API
  version: 1.0.0
  description: Modular, cryptographically secure blockchain-based data management core for cross-module integration and strategic decision auditing.
servers:
  - url: http://127.0.0.1:3000
    description: Local Development Server
paths:
...
  /api/module/sync:
    post:
      summary: Synchronize data from an external module
      description: Accepts payloads tagged with a specific source module identifier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sourceModule
                - payload
              properties:
                sourceModule:
                  type: string
                  example: AdvisoryBoard
                payload:
                  type: string
                  example: Approved technological independence strategy
      responses:
        "200":
          description: Module data synchronized into a new block
...